test(sim): split the second-tier balance suites and rebalance the lanes
Splits the chronomancy suite three ways along its describe boundaries
(the shared rotation harness moved verbatim to tests/helpers/, since the
measurement blocks run at collection time) and the warlock sustain suite
per spec, so the last multi-minute single files leave the shard pool's
worst tails. druid_balance_probe stays whole: its matrix assertions are
an argmax across capstones and a split would weaken the winner pin.

Rebalances the lane halves from the run 31732244215 per-file ledger:
battleground (14.4s) and audit_conservation_property (55.1s) leave for
the shard pool per the 90 second rule, only the chronomancy targets file
stays lane-listed of its split, and half b is now derived from the union
minus half a so an unowned lane file is structurally impossible.

Also carries the review round on the harness split: a literal pin for
both arms of the diet helper, the lane-ownership rule exception stated
where the rule lives, repointed doc anchors, and the healer-profiles
timeout re-sized for its new cold-start position.
