+++ scripts/lib/ci_shard_plan.mjs
@@ -71,9 +71,14 @@ export const CI_GUARD_PREFIXES = Object.freeze(['tests/parity/']);
  * single multi-minute file stops setting the slowest shard's wall clock.
  * Membership is measured, not automated: a file joins when it costs more than
  * the 90 second threshold inside a full-mode CI shard, and the next-longest
- * file stays sharded (2026-08-06 full-mode run: these four measured 249.5 s,
- * 143.9 s, 142.3 s, and 94.7 s in-shard; the next longest,
- * tests/corpse_harvest_sim.test.ts at 69.8 s, stays). Membership is also
+ * file stays sharded (2026-08-06 full-mode run: the then-four members
+ * measured 249.5 s, 143.9 s, 142.3 s, and 94.7 s in-shard; the next longest,
+ * tests/corpse_harvest_sim.test.ts at 69.8 s, stays). One decided exception
+ * to the per-file rule: the owned-class balance family split 2026-08-13 is
+ * lane-owned as a UNIT, split pieces included, so the diet flag registry
+ * (every WOC_FULL_BALANCE_SWEEP reader must be lane-owned) and the family's
+ * lane accounting stay in one place; measured per-file lane durations for
+ * the whole family are in the split PR (#3370). Membership is also
  * one-directional by nature: nothing automatic promotes a newly slowed suite
  * into this list, so when a shard's wall clock grows, remeasure from the
  * shard logs and re-decide the list (the committed contract and the audit
@@ -83,9 +88,16 @@ export const CI_GUARD_PREFIXES = Object.freeze(['tests/parity/']);
  * whole suite in their 8 shards, so the post-merge backstop is untouched.
  */
 export const CI_LONG_SUITES = Object.freeze([
-  'tests/audit_conservation_property.test.ts',
-  'tests/battleground.test.ts',
-  'tests/chronomancy_balance.test.ts',
+  // 2026-08-13 remeasure (run 31732244215, both lanes fully loaded):
+  // battleground (14.4 s) and audit_conservation_property (55.1 s) had drifted
+  // far under the 90 second rule and left for the shard pool; the chronomancy
+  // suite split three ways and only its balance-targets file still clears the
+  // rule (the heal-parity and Cascada pieces shard); the warlock sustain suite
+  // split per spec and its pieces shard pending an in-shard measurement.
+  // druid_balance_probe stays WHOLE: its cost is one matrix test whose
+  // bestDruidBuilds assertions are an argmax across capstones, so a
+  // per-capstone split would weaken the winner selection it pins.
+  'tests/chronomancy_balance_targets.test.ts',
   // The five-class-overhauls balance harnesses (review 3050): the owned-class
   // matrices grew to 8 specs and the raid loop to ~510s, pushing shards 1 and
   // 4 past the then-20-minute pr-gate shard budget; they are exactly what this
@@ -124,27 +136,23 @@ export const CI_LONG_SUITES = Object.freeze([
  * without touching the shard side. tests/ci_shard_plan.test.ts pins the
  * halves as an exact partition of CI_LONG_SUITES.
  */
+const CI_LONG_SUITE_HALF_A = Object.freeze([
+  'tests/nythraxis_matrix.test.ts',
+  'tests/owned_class_balance_dps_metrics.test.ts',
+  'tests/owned_class_balance_dps_probes.test.ts',
+  'tests/owned_class_balance_druid_bands.test.ts',
+  'tests/owned_class_balance_healer_probes.test.ts',
+  'tests/owned_class_balance_role_bands.test.ts',
+  'tests/owned_class_raid_armor_avoidance.test.ts',
+]);
+
+// Half b is DERIVED (the union minus half a), so a lane file excluded from
+// every shard but owned by no lane is structurally impossible, not merely
+// pinned; tests/ci_shard_plan.test.ts still pins b's exact contents so the
+// derived assignment stays a conscious decision.
 export const CI_LONG_SUITE_HALVES = Object.freeze({
-  a: Object.freeze([
-    'tests/battleground.test.ts',
-    'tests/chronomancy_balance.test.ts',
-    'tests/druid_balance_probe.test.ts',
-    'tests/owned_class_balance_dps_probes.test.ts',
-    'tests/owned_class_balance_groveheart.test.ts',
-    'tests/owned_class_balance_role_bands.test.ts',
-    'tests/owned_class_raid_armor_avoidance.test.ts',
-  ]),
-  b: Object.freeze([
-    'tests/audit_conservation_property.test.ts',
-    'tests/eastbrook_gameplay_integration.test.ts',
-    'tests/hunter_dps_balance.test.ts',
-    'tests/nythraxis_matrix.test.ts',
-    'tests/owned_class_balance_dps_metrics.test.ts',
-    'tests/owned_class_balance_druid_bands.test.ts',
-    'tests/owned_class_balance_healer_contract.test.ts',
-    'tests/owned_class_balance_healer_probes.test.ts',
-    'tests/owned_class_raid_sustain_bands.test.ts',
-  ]),
+  a: CI_LONG_SUITE_HALF_A,
+  b: Object.freeze(CI_LONG_SUITES.filter((f) => !CI_LONG_SUITE_HALF_A.includes(f))),
 });
 
 /**
