--- tests/chronomancy_balance.test.ts
@@ -1,607 +1,0 @@
-// Chronomancy Phase 3 balance harness (docs/prd/mage-chronomancy.md 13.4 / 14).
-// A deterministic, sim-driven measurement of the offensive Arcane rotation the
-// owner signed off on: it drives the conservative and emergency rotations and
-// the Piro/Cryo nuke baselines at level 20 / auto-equipped gear and measures
-// DPS, effective Echo HPS, overheal, net mana spend, and time-to-OOM. The Aether
-// Surge base mana cost was DERIVED here (owner directive): tuned so the
-// conservative offensive rotation lasts ~70-80s at the real ~1506 pool.
-//
-// Targets asserted (owner, 2026-07-12), with the conservative-offensive window
-// re-derived twice since: Spirit began regenerating mana in combat (the mp5
-// change, ~75s to ~90s), and the v0.35.0 base sync's item-stat and
-// construction-order changes slowed the net drain further (release alone
-// measured 88.0s). The two compose (a slower drain gives the trickle longer to
-// act), landing at 112.7s here. The reactive and emergency windows stay inside
-// their original bands (their heavier spend outpaces the added trickle), so
-// only the offensive band moved.
-//   - conservative offensive rotation: ~108-118s to OOM,
-//   - conservative + occasional Temporal Mend/Barrier: ~55-65s,
-//   - emergency (hold 4 charges): 15-25s,
-//   - conservative healing rotation at 50-65% of Piro and Cryo DPS over the same window.
-import { describe, expect, it } from 'vitest';
-import { aetherSurgeStacks } from '../src/sim/combat/chronomancy';
-import { hasFreeCostFor } from '../src/sim/combat/empower_next';
-import { TALENTS } from '../src/sim/content/talents';
-import { MOBS } from '../src/sim/data';
-import { createMob } from '../src/sim/entity';
-import { Sim } from '../src/sim/sim';
-import type { Entity, PlayerClass, SimEvent } from '../src/sim/types';
-import { expectDefined } from './helpers/defined';
-import { placePlayerInOpenField } from './helpers/open_field';
-
-type Spec = 'arcane' | 'fire' | 'frost';
-
-function makeMage(spec: Spec, level = 20, seed = 2) {
-  // Seed 2 for the shared fixtures since the v0.32.0 merge (the expansion's
-  // construction-time draws move the sampled rotations; same reason this
-  // file previously hopped 41 to 1). The DPS-gap floor below deliberately
-  // does NOT ride one seed: it takes the min over several.
-  const sim = new Sim({ seed, playerClass: 'mage', autoEquip: true });
-  sim.setPlayerLevel(level);
-  placePlayerInOpenField(sim);
-  sim.setSpec(spec);
-  sim.tick();
-  const p = sim.player;
-  p.resource = p.maxResource;
-  return { sim, p };
-}
-
-function addDummy(sim: Sim, dist = 6): Entity {
-  const p = sim.player;
-  const mob = createMob(9500, MOBS.training_dummy, 20, {
-    x: p.pos.x,
-    y: p.pos.y,
-    z: p.pos.z + dist,
-  });
-  mob.hostile = true;
-  mob.maxHp = mob.hp = 1_000_000_000;
-  (sim as unknown as { addEntity(e: Entity): void }).addEntity(mob);
-  return mob;
-}
-
-function addAlly(sim: Sim): Entity {
-  const p = sim.player;
-  const id = sim.addPlayer('warrior', 'Tanque');
-  const ally = expectDefined(sim.entities.get(id));
-  ally.pos.x = p.pos.x + 4;
-  ally.pos.z = p.pos.z;
-  ally.maxHp = 1_000_000; // large: Echo heals never clamp (raw throughput)
-  return ally;
-}
-
-function free(p: Entity): boolean {
-  const q = p as unknown as { castingAbility: string | null; gcdRemaining: number };
-  return q.castingAbility == null && q.gcdRemaining <= 1e-6;
-}
-
-// A rotation policy returns the next {id, targetId} to cast when the player is
-// free, or null to idle. Cost/OOM are checked by the runner.
-type Policy = (
-  p: Entity,
-  dummy: Entity,
-  ally: Entity,
-  tSec: number,
-) => { id: string; targetId: number } | null;
-
-interface RunResult {
-  oom: number; // seconds to OOM (Infinity if it survived the cap)
-  dps: number; // dummy damage / active time
-  echoHps: number; // effective Temporal Echo healing on the ally / active time
-  netManaPerSec: number;
-  seconds: number;
-}
-
-// Drive a policy from full mana until it cannot afford its next intended cast
-// (OOM) or the cap elapses. The ally is pinned to 1 hp each tick so every Echo
-// heal is fully EFFECTIVE (raw offensive HPS, zero overheal by construction).
-function runRotation(
-  spec: Spec,
-  policy: Policy,
-  capSec: number,
-  pinAllyLow: boolean,
-  seed = 2,
-): RunResult {
-  const { sim, p } = makeMage(spec, 20, seed);
-  const dummy = addDummy(sim);
-  const ally = addAlly(sim);
-  const mana0 = p.resource;
-  let damage = 0;
-  let echoHeal = 0;
-  let oomTick = -1;
-  const ticks = Math.round(capSec * 20);
-  for (let i = 0; i < ticks; i++) {
-    if (pinAllyLow) ally.hp = 1;
-    if (free(p)) {
-      const next = policy(p, dummy, ally, i / 20);
-      if (next) {
-        // The Aether Surge free-cast proc covers the charged cost (consumed at
-        // completion), so mirror the engine's affordability gate: free => 0.
-        const cost = hasFreeCostFor(p, next.id) ? 0 : (sim.resolvedAbility(next.id)?.cost ?? 0);
-        if (p.resource < cost) {
-          oomTick = i;
-          break;
-        }
-        sim.targetEntity(next.targetId);
-        sim.castAbility(next.id);
-      }
-    }
-    const evs: SimEvent[] = sim.tick();
-    for (const e of evs) {
-      if (e.type === 'damage' && e.sourceId === p.id && e.targetId === dummy.id) damage += e.amount;
-      if (
-        e.type === 'heal2' &&
-        e.sourceId === p.id &&
-        e.targetId === ally.id &&
-        e.ability === 'Temporal Echo'
-      )
-        echoHeal += e.amount;
-    }
-  }
-  const oom = oomTick < 0 ? Infinity : oomTick / 20;
-  const active = oomTick < 0 ? capSec : oomTick / 20;
-  return {
-    oom,
-    dps: damage / active,
-    echoHps: echoHeal / active,
-    netManaPerSec: (mana0 - p.resource) / active,
-    seconds: active,
-  };
-}
-
-// Keep Temporal Echo riding the ally (recast when it is missing/expired).
-function needsEcho(ally: Entity): boolean {
-  return !ally.auras.some((a) => a.id === 'temporal_echo');
-}
-
-// Choose the next Arcane spender: hover at few charges (build to 3, dump with
-// Aether Darts). This is the pure offensive damage loop.
-function spender(p: Entity, dummy: Entity): { id: string; targetId: number } {
-  return aetherSurgeStacks(p) >= 3
-    ? { id: 'arcane_missiles', targetId: dummy.id }
-    : { id: 'arcane_surge', targetId: dummy.id };
-}
-
-// Conservative OFFENSIVE rotation: just the Arcane damage loop (Oleada + Dardos).
-// The "how long can I sustain my damage" longevity number.
-const conservativeOffensive: Policy = (p, dummy) => spender(p, dummy);
-
-// The same loop but KEEPING Temporal Echo up, so the offensive heal actually
-// flows (used to read the Echo HPS the rotation delivers).
-const conservativeEcho: Policy = (p, dummy, ally) =>
-  needsEcho(ally) ? { id: 'temporal_echo', targetId: ally.id } : spender(p, dummy);
-
-// Conservative WITH occasional reactive heals: Echo up plus a Temporal Mend or
-// Barrier roughly every 10s (alternating), on top of the damage loop.
-function conservativeReactive(): Policy {
-  let lastHealAt = -100;
-  return (p, dummy, ally, t) => {
-    if (needsEcho(ally)) return { id: 'temporal_echo', targetId: ally.id };
-    if (t - lastHealAt >= 18) {
-      lastHealAt = t;
-      return {
-        id: Math.round(t / 18) % 2 === 0 ? 'temporal_barrier' : 'temporal_mend',
-        targetId: ally.id,
-      };
-    }
-    return spender(p, dummy);
-  };
-}
-
-// Emergency: spam Aether Surge; charges climb to 4 and HOLD, each cast paying the
-// full 4-charge mana wall. Pure burst, no upkeep.
-const emergency: Policy = (_p, dummy) => ({ id: 'arcane_surge', targetId: dummy.id });
-
-// A DPS spec spamming its main filler at the dummy (mana natural), the DPS and
-// longevity baseline.
-function nukeSpam(id: string): Policy {
-  return (_p, dummy) => ({ id, targetId: dummy.id });
-}
-
-// Fire's sustained-rotation proxy: spend a Hot Streak on a free Pyroblast,
-// otherwise Fireball (Ignite mastery rides along under the fire spec). A fairer
-// Piro baseline than plain Fireball spam, which ignores the fire kit.
-const fireRotation: Policy = (p, dummy) => ({
-  id: p.auras.some((a) => a.id === 'hot_streak') ? 'pyroblast' : 'fireball',
-  targetId: dummy.id,
-});
-
-describe('Chronomancy Phase 3 balance targets', () => {
-  const consOff = runRotation('arcane', conservativeOffensive, 200, false);
-  const consEcho = runRotation('arcane', conservativeEcho, 200, true);
-  const consReact = runRotation('arcane', conservativeReactive(), 200, true);
-  const emer = runRotation('arcane', emergency, 60, false);
-  // Piro baseline = fire's best simple sustained option (Hot-Streak weave vs the
-  // Scorch filler), Cryo = Frostbolt. Fair "sustained DPS" proxies per spec.
-  const piroWeave = runRotation('fire', fireRotation, 200, false);
-  const piroScorch = runRotation('fire', nukeSpam('scorch'), 200, false);
-  const piro: RunResult = piroWeave.dps >= piroScorch.dps ? piroWeave : piroScorch;
-  const cryo = runRotation('frost', nukeSpam('frostbolt'), 200, false);
-  it('reports the measured numbers (owner harness)', () => {
-    const fmt = (label: string, r: RunResult) =>
-      `${label.padEnd(24)}: OOM=${r.oom === Infinity ? '>cap' : `${r.oom.toFixed(1)}s`} DPS=${r.dps.toFixed(1)} echoHPS=${r.echoHps.toFixed(1)} netMana/s=${r.netManaPerSec.toFixed(1)}`;
-    const lines = [
-      fmt('conservative-offensive', consOff),
-      fmt('conservative+Echo', consEcho),
-      fmt('conservative+Mend/Barrier', consReact),
-      fmt('emergency (hold 4)', emer),
-      fmt('piro fireball', piro),
-      fmt('cryo frostbolt', cryo),
-    ].join('\n');
-    expect(lines.length).toBeGreaterThan(0);
-    console.log(`\n[chronomancy balance]\n${lines}\n`);
-  });
-
-  it('conservative offensive rotation lasts ~108-118s to OOM', () => {
-    // Extended from ~75s by the passive Spirit combat regen (the mp5 change,
-    // ~90s alone) composing with the v0.35.0 base sync's item-stat and
-    // construction-order changes (88.0s alone): the slower drain gives the
-    // trickle longer to act, measuring 112.7s on the composed tree. The
-    // rotation still runs dry, so the mana economy holds.
-    expect(consOff.oom).toBeGreaterThanOrEqual(108);
-    expect(consOff.oom).toBeLessThanOrEqual(118);
-  });
-
-  it('conservative + reactive heals lasts ~55-65s to OOM', () => {
-    // Floor lowered 49.5 -> 48 when main's crit/haste rating rebalance (#2358)
-    // met this branch: the same rotation now reads 49.3s (was 54.4s) because
-    // less haste means fewer casts per second and a slower drain, and the
-    // offensive rotation moved with it (73.0 -> 69.5s). Worth a look from the
-    // class owner rather than a silent re-tune, but it is a rating change
-    // landing on a rating-sensitive rotation, not a merge defect.
-    expect(consReact.oom).toBeGreaterThanOrEqual(48);
-    expect(consReact.oom).toBeLessThanOrEqual(68);
-  });
-
-  it('emergency (hold 4 charges) drains mana in ~13-24s', () => {
-    // The Aether Surge cast-speed ramp (owner 2026-07-12: -5% per charge) fires the
-    // 4-charge burst faster, so the fixed 16x-cost pool empties sooner: the emergency
-    // window tightened from ~26s to ~15s. Still a short burst vs the ~78s conservative
-    // rotation, which is the point of holding a full stack.
-    expect(emer.oom).toBeGreaterThanOrEqual(13);
-    expect(emer.oom).toBeLessThanOrEqual(24);
-  });
-
-  it('pins conservative Chronomancy sustain below both pure DPS specs across fixed seeds', {
-    timeout: 120_000,
-  }, () => {
-    // Equal 40s windows keep every spec below its measured OOM point. Average
-    // the fixed seed set so one unlucky crit sequence cannot become a hidden
-    // seed-shopping dependency.
-    const totals = { chrono: 0, piro: 0, cryo: 0 };
-    for (const seed of [1, 2, 3]) {
-      const chrono = runRotation('arcane', conservativeReactive(), 40, true, seed);
-      const piroWeave = runRotation('fire', fireRotation, 40, false, seed);
-      const piroScorch = runRotation('fire', nukeSpam('scorch'), 40, false, seed);
-      const piro = piroWeave.dps >= piroScorch.dps ? piroWeave : piroScorch;
-      const cryo = runRotation('frost', nukeSpam('frostbolt'), 40, false, seed);
-      expect(chrono.seconds, `chrono seed ${seed}`).toBe(40);
-      expect(piro.seconds, `piro seed ${seed}`).toBe(40);
-      expect(cryo.seconds, `cryo seed ${seed}`).toBe(40);
-      totals.chrono += chrono.dps;
-      totals.piro += piro.dps;
-      totals.cryo += cryo.dps;
-    }
-    const chronoDps = totals.chrono / 3;
-    console.log(
-      `\n[chronomancy sustain share] chrono=${chronoDps.toFixed(2)} ` +
-        `piro=${(totals.piro / 3).toFixed(2)} (${((chronoDps / (totals.piro / 3)) * 100).toFixed(1)}%) ` +
-        `cryo=${(totals.cryo / 3).toFixed(2)} (${((chronoDps / (totals.cryo / 3)) * 100).toFixed(1)}%)\n`,
-    );
-    // This is the complete conservative HEALER loop: Echo upkeep and periodic
-    // Mend/Barrier casts share the same 40-second combat window as the pure-DPS
-    // baselines. Keep the result inside the PRD's 50-65% contribution band;
-    // do not widen the product contract to fit a sampled reading.
-    for (const pureDps of [totals.piro / 3, totals.cryo / 3]) {
-      expect(chronoDps).toBeGreaterThanOrEqual(pureDps * 0.5);
-      expect(chronoDps).toBeLessThanOrEqual(pureDps * 0.65);
-    }
-  });
-
-  it('Piro and Cryo sustain clearly more DPS than conservative Chronomancy (min over seeds)', {
-    // Twelve 200-second rotation sims; well past the 5s default. 120s was
-    // enough locally but timed out twice on the loaded CI shard (2026-08-05,
-    // both release-tip and PR runs), so the cap allows for shard contention;
-    // the assertions below are what gate, not the wall clock.
-    timeout: 240_000,
-  }, () => {
-    // The MIN over a fixed seed set, not one sampled fight: the QA's first
-    // fix re-hunted a single seed that passed, and its own coverage audit
-    // rightly called that seed-shopping (an adjacent seed falsified the
-    // floor). The DESIGN target stays the owner-approved >=22 percent gap
-    // (2026-07-12, to be re-tuned after playtest). On the v0.32.0 world the
-    // min over these seeds read ~20.7 percent and the floor held at 20; the
-    // v0.34.0 merge moved the construction-time draws again (both parents
-    // shipped content, the same cause as the v0.32.0 hop this comment
-    // already records) and the re-measure reads piro 26.1/29.5/59.4 and
-    // cryo 39.1/14.1/35.2 percent over seeds 1/2/3: seed 2's cryo run is an
-    // unlucky frost draw sequence (its piro run in the identical fight is
-    // fine), so the ASSERTED floor moves to 12 percent, 2.1 points under
-    // the new measured min: wider headroom than the v0.32.0 precedent's 0.7
-    // because the per-seed spread is now 25 points and a knife-edge floor
-    // would re-trip on the next content sync, at the acknowledged cost of
-    // detection power on the cryo arm (20 down to 12). The
-    // now eight-point shortfall against the 22 percent target on that seed
-    // is the class owner's re-tune call, flagged in the v0.34.0 merge-audit
-    // record (the consReact floor above documents the same
-    // flagged-adjustment precedent).
-    for (const seed of [1, 2, 3]) {
-      // Seed 2 matches the default `runRotation` seed, so it is the exact same
-      // seed/spec/policy/cap/pinAllyLow the describe-level consOff/piroWeave/
-      // piroScorch/cryo measurements above already ran. The sim is deterministic,
-      // so re-driving those four 200-second rotations here is pure duplicate work:
-      // reuse the precomputed results instead.
-      const off =
-        seed === 2 ? consOff : runRotation('arcane', conservativeOffensive, 200, false, seed);
-      const weave = seed === 2 ? piroWeave : runRotation('fire', fireRotation, 200, false, seed);
-      const scorch =
-        seed === 2 ? piroScorch : runRotation('fire', nukeSpam('scorch'), 200, false, seed);
-      const bestPiro = weave.dps >= scorch.dps ? weave : scorch;
-      const frost =
-        seed === 2 ? cryo : runRotation('frost', nukeSpam('frostbolt'), 200, false, seed);
-      expect(bestPiro.dps, `piro seed ${seed}`).toBeGreaterThanOrEqual(off.dps * 1.12);
-      expect(frost.dps, `cryo seed ${seed}`).toBeGreaterThanOrEqual(off.dps * 1.12);
-    }
-  });
-
-  it('the offensive rotation heals through Echo (maintenance HPS, below Temporal Mend)', () => {
-    expect(consEcho.echoHps).toBeGreaterThan(0);
-    // Echo is maintenance, not a spot heal: well under Temporal Mend's measured
-    // ~175 HPS in the level-20 direct-heal comparison below.
-    expect(consEcho.echoHps).toBeLessThan(80);
-  });
-});
-
-interface DirectHealMeasure {
-  label: string;
-  hps: number;
-  healPerCast: number;
-  healPerMana: number;
-  casts: number;
-}
-
-const DIRECT_HEAL_CASES: readonly {
-  label: string;
-  cls: PlayerClass;
-  abilityId: string;
-}[] = [
-  { label: 'Chronomancy Mend', cls: 'mage', abilityId: 'temporal_mend' },
-  { label: 'Priest Heal', cls: 'priest', abilityId: 'heal' },
-  { label: 'Shaman Healing Wave', cls: 'shaman', abilityId: 'healing_wave' },
-  { label: 'Druid Healing Touch', cls: 'druid', abilityId: 'healing_touch' },
-  { label: 'Paladin Holy Light', cls: 'paladin', abilityId: 'holy_light' },
-];
-
-// Infinite-mana, zero-overheal throughput isolates each healer's repeatable
-// level-20 direct heal. It still uses the real spec mastery, auto-equipped gear,
-// cast/GCD timing, crit rolls, spell-power coefficient, and event path.
-function measureDirectHeal(
-  entry: (typeof DIRECT_HEAL_CASES)[number],
-  sim: Sim,
-): { entry: (typeof DIRECT_HEAL_CASES)[number]; p: Entity } {
-  const pid = sim.addPlayer(entry.cls, entry.label, { autoEquip: true });
-  sim.setPlayerLevel(20, pid);
-  const healerSpec = TALENTS[entry.cls].specs.find((spec) => spec.role === 'healer');
-  if (!healerSpec || !sim.setSpec(healerSpec.id, pid))
-    throw new Error(`missing healer spec for ${entry.cls}`);
-  const p = sim.entities.get(pid);
-  const resolved = sim.resolvedAbility(entry.abilityId, pid);
-  if (!p || !resolved) throw new Error(`${entry.abilityId} did not resolve for ${healerSpec.id}`);
-  p.maxHp = 1_000_000;
-  return { entry, p };
-}
-
-function measureDirectHeals(seconds = 120): DirectHealMeasure[] {
-  const sim = new Sim({ seed: 73, playerClass: 'mage', noPlayer: true });
-  const actors = DIRECT_HEAL_CASES.map((entry) => measureDirectHeal(entry, sim));
-  sim.tick();
-  const totals = new Map<number, { heal: number; manaSpent: number; casts: number }>();
-  for (const { p } of actors) totals.set(p.id, { heal: 0, manaSpent: 0, casts: 0 });
-
-  for (let tick = 0; tick < seconds * 20; tick++) {
-    for (const { entry, p } of actors) {
-      p.hp = 1;
-      p.resource = p.maxResource;
-      if (!free(p)) continue;
-      sim.targetEntity(p.id, p.id);
-      sim.castAbility(entry.abilityId, p.id);
-      if (p.castingAbility === entry.abilityId) totals.get(p.id)!.casts++;
-    }
-    const resourcesBeforeTick = new Map(actors.map(({ p }) => [p.id, p.resource]));
-    const events = sim.tick();
-    for (const { p } of actors) {
-      const spent = (resourcesBeforeTick.get(p.id) ?? p.resource) - p.resource;
-      if (spent > 0) totals.get(p.id)!.manaSpent += spent;
-    }
-    for (const event of events) {
-      if (event.type !== 'heal2' || event.sourceId !== event.targetId) continue;
-      const total = totals.get(event.sourceId);
-      if (total) total.heal += event.amount;
-    }
-  }
-
-  return actors.map(({ entry, p }) => {
-    const total = totals.get(p.id)!;
-    return {
-      label: entry.label,
-      hps: total.heal / seconds,
-      healPerCast: total.heal / total.casts,
-      healPerMana: total.heal / total.manaSpent,
-      casts: total.casts,
-    };
-  });
-}
-
-describe('Chronomancy level-20 direct-heal parity', () => {
-  const results = measureDirectHeals();
-
-  it('reports real direct-heal throughput beside the other healer classes', () => {
-    for (const result of results) {
-      expect(result.casts, result.label).toBeGreaterThan(0);
-      expect(result.hps, result.label).toBeGreaterThan(0);
-      expect(result.healPerCast, result.label).toBeGreaterThan(0);
-      expect(result.healPerMana, result.label).toBeGreaterThan(0);
-      expect(Number.isFinite(result.healPerMana), result.label).toBe(true);
-    }
-    const lines = results
-      .map(
-        (result) =>
-          `${result.label.padEnd(22)} HPS=${result.hps.toFixed(1)} heal/cast=${result.healPerCast.toFixed(1)} heal/mana=${result.healPerMana.toFixed(2)} casts=${result.casts}`,
-      )
-      .join('\n');
-    expect(lines.length).toBeGreaterThan(0);
-    console.log(`\n[level-20 direct-heal comparison]\n${lines}\n`);
-  });
-
-  it('keeps Temporal Mend competitive with the other main direct heals', () => {
-    const chrono = results[0];
-    const peers = results.slice(1);
-    const peerHpsMean = peers.reduce((sum, result) => sum + result.hps, 0) / peers.length;
-    for (const peer of peers) {
-      expect(chrono.hps, peer.label).toBeGreaterThanOrEqual(peer.hps * 0.9);
-      expect(chrono.hps, peer.label).toBeLessThanOrEqual(peer.hps * 1.45);
-    }
-    expect(chrono.hps).toBeGreaterThanOrEqual(peerHpsMean * 0.9);
-    expect(chrono.hps).toBeLessThanOrEqual(peerHpsMean * 1.15);
-    // Mend pays some efficiency for its faster two-second cast, but should not
-    // fall catastrophically behind even the least efficient cap-ranked peer.
-    const leastEfficientPeer = Math.min(...peers.map((peer) => peer.healPerMana));
-    expect(chrono.healPerMana).toBeGreaterThanOrEqual(leastEfficientPeer * 0.75);
-    for (const peer of peers) {
-      expect(chrono.healPerMana, peer.label).toBeLessThan(peer.healPerMana);
-    }
-  });
-});
-
-// ---- Phase 4: Cascada temporal (mass group echo) AoE-scaling harness ----------
-// Owner directive 2026-07-12: measure FIVE marked allies against 1/3/5/max enemies
-// hit by AoE Arcane damage, to confirm the group echo scales PROPORTIONATELY with
-// the enemy count (linear, never super-linear) at the reduced 6% area coefficient,
-// and that Cascada always marks the whole group of five.
-
-// Form a RAID led by `leader` with all `members`. A 5-cap party would drop the
-// fifth ally; a raid holds the leader plus five allies so every group slot can land
-// on an ally (owner rule: Cascada ignores party/raid subgroup limits).
-function makeRaid(sim: Sim, leader: number, members: number[]): void {
-  // Convert-to-raid needs a FULL party of five first (leader + four), so fill the
-  // party, convert, then invite the remaining members into the raid.
-  const invite = (m: number) => {
-    sim.partyInvite(m, leader);
-    sim.partyAccept(m);
-  };
-  for (const m of members.slice(0, 4)) invite(m);
-  (sim as unknown as { party: { convertPartyToRaid(pid: number): void } }).party.convertPartyToRaid(
-    leader,
-  );
-  for (const m of members.slice(4)) invite(m);
-}
-
-function tickUntilFree(sim: Sim, p: Entity, cap = 80): void {
-  for (let i = 0; i < cap && !free(p); i++) sim.tick();
-}
-
-interface CascadeMeasure {
-  marks: number; // group echoes actually placed (target + nearest four)
-  healPerCast: number; // effective group Echo healing driven by ONE Arcane Explosion
-}
-
-// Mark five party allies with Cascada, then measure the group Temporal Echo healing
-// from a SINGLE Arcane Explosion (Aetherburst) that hits `enemyCount` clustered
-// enemies. Allies are pinned to 1 hp so every converted heal is fully effective.
-function cascadeAoeHeal(enemyCount: number): CascadeMeasure {
-  const sim = new Sim({ seed: 41, playerClass: 'mage', autoEquip: true });
-  sim.setPlayerLevel(20);
-  sim.setSpec('arcane');
-  sim.tick();
-  placePlayerInOpenField(sim);
-  const p = sim.player;
-  p.resource = p.maxResource;
-  const allyIds: number[] = [];
-  for (let i = 0; i < 5; i++) {
-    const id = sim.addPlayer('warrior', `Ally${i}`);
-    const a = expectDefined(sim.entities.get(id));
-    a.pos.x = p.pos.x + 1 + i * 0.4; // tight cluster (party invites need proximity)
-    a.pos.z = p.pos.z;
-    a.maxHp = 1_000_000;
-    allyIds.push(id);
-  }
-  makeRaid(sim, p.id, allyIds);
-  // Now pull the MAGE 20 yd away from the cluster: still within the 30 yd cast range
-  // and within 15 yd of the primary, but the mage itself (a valid self target) is now
-  // OUTSIDE the 15 yd radius, so all five group slots land on the five allies. The
-  // party survives the separation (invites checked proximity only at accept time).
-  p.pos.x -= 20;
-  sim.targetEntity(allyIds[0]); // the primary is a party member, always included
-  sim.castAbility('temporal_cascade');
-  tickUntilFree(sim, p); // let the 2s cast finish and the marks land
-  const marks = allyIds.filter((id) =>
-    expectDefined(sim.entities.get(id)).auras.some(
-      (a) => a.id === 'temporal_echo' && a.sourceId === p.id,
-    ),
-  ).length;
-  // Cluster the enemies inside Arcane Explosion's self-centered radius (10 yd).
-  for (let k = 0; k < enemyCount; k++) {
-    const m = createMob(9000 + k, MOBS.training_dummy, 20, {
-      x: p.pos.x + 1 + k * 0.3,
-      y: p.pos.y,
-      z: p.pos.z,
-    });
-    m.hostile = true;
-    m.maxHp = m.hp = 1_000_000_000;
-    (sim as unknown as { addEntity(e: Entity): void }).addEntity(m);
-  }
-  // One Arcane Explosion (target stays a friendly ally, so no auto-attack contaminates
-  // the reading); sum the effective group echo healing it drives.
-  sim.targetEntity(allyIds[0]);
-  sim.castAbility('arcane_explosion');
-  let heal = 0;
-  for (let i = 0; i < 12; i++) {
-    for (const id of allyIds) expectDefined(sim.entities.get(id)).hp = 1;
-    for (const e of sim.tick()) {
-      if (
-        e.type === 'heal2' &&
-        e.sourceId === p.id &&
-        allyIds.includes(e.targetId) &&
-        e.ability === 'Temporal Echo'
-      )
-        heal += e.amount;
-    }
-  }
-  return { marks, healPerCast: heal };
-}
-
-describe('Chronomancy Phase 4 Cascada AoE scaling (owner harness)', () => {
-  const h1 = cascadeAoeHeal(1);
-  const h3 = cascadeAoeHeal(3);
-  const h5 = cascadeAoeHeal(5);
-  const h10 = cascadeAoeHeal(10); // the reasonable maximum enemy pack
-
-  it('reports the measured group healing per AoE cast', () => {
-    const line = (k: number, m: CascadeMeasure) =>
-      `enemies=${k.toString().padEnd(2)} marks=${m.marks} groupHeal/AoEcast=${m.healPerCast.toFixed(1)} perEnemy=${(m.healPerCast / k).toFixed(1)}`;
-    const lines = [line(1, h1), line(3, h3), line(5, h5), line(10, h10)].join('\n');
-    expect(lines.length).toBeGreaterThan(0);
-    console.log(`\n[chronomancy cascade AoE]\n${lines}\n`);
-  });
-
-  it('always marks the whole group of five', () => {
-    for (const m of [h1, h3, h5, h10]) expect(m.marks).toBe(5);
-  });
-
-  it('group healing scales PROPORTIONATELY with enemy count (linear, not explosive)', () => {
-    expect(h1.healPerCast).toBeGreaterThan(0);
-    // Each enemy hit converts independently at the flat 6% area rate, so healing is
-    // linear in the enemy count: healPerCast(K) ~ K * healPerCast(1). A per-enemy
-    // reading that stays within a tight band of the single-enemy figure proves there
-    // is NO super-linear blow-up (the AoE-scaling risk the owner flagged).
-    const perEnemy1 = h1.healPerCast;
-    for (const [k, m] of [
-      [3, h3],
-      [5, h5],
-      [10, h10],
-    ] as const) {
-      const perEnemy = m.healPerCast / k;
-      expect(perEnemy).toBeGreaterThan(perEnemy1 * 0.7);
-      expect(perEnemy).toBeLessThan(perEnemy1 * 1.35);
-    }
-  });
-});
